provisioner{,_cli}: Send build events using a channel
Why not a stream? For methods that return things, such as a git clone,
we can't return a stream since the primary output is something else
(namely, a tarball.) Therefore we can emit output on the side if the
caller passes in an (optional!) sender to emit events on. Senders can be
cloned anyways, cheaply, so there is no issue there regarding
allocations.